refactor(core): split agent_task_promotion god-file into module dir
Splits src/core/agent_task_promotion.rs (1732 lines, 48 items — baselined
GodFile + HighItemCount) into an agent_task_promotion/ module dir (types,
promote, apply) with the test module moved to tests.rs, each under
structural thresholds, minimal re-exporting mod.rs. Pure move; baseline
rows removed.
